Init commit

+++ src/main/java/jobportal/repository/JobSeekerProfileRepository.java
@@ -3,5 +3,5 @@ package jobportal.repository;
 import jobportal.entity.JobSeekerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobSeekerProfileRepository extends JpaRepository<JobSeekerProfile, Long> {
+public interface JobSeekerProfileRepository extends JpaRepository<JobSeekerProfile, Integer> {
 }
